refactor: removed duplicate code (checkbox) closes #262
